requirements.txt
base58==2.1.1
Flask==3.1.0
GitPython==3.1.44
PyJWT==2.10.1
pyotp==2.9.0
Requests==2.32.3
